pyproject.toml
[tool.poetry]
name = "news-tracker"
version = "0.1.0"
description = ""
authors = ["Marek Majdak <22301547+marekm1844@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pydantic = "^2.9.2"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
newspaper3k = "^0.2.8"
lxml-html-clean = "^0.3.1"
beautifulsoup4 = "^4.12.3"
simplediff = "^1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
